+++ src/core.ts
@@ -1,10 +1,9 @@
 import * as path from "path";
-import * as config from "./config";
 import { workspace, Disposable } from 'vscode';
 import { AVDService } from './service/AVDService';
 import { SDKService } from './service/SDKService';
 import { AndroidService } from './service/AndroidService';
-import * as Module from "./module";
+
 import { Output } from "./module/ui";
 import { Cache } from "./module/cache";
 import { Platform } from "./module/platform";
@@ -27,8 +26,22 @@ export interface IConfig {
     emulatorOpt?: string
 }
 
+export enum ConfigItem {
+    sdkPath = "sdkPath",
+    cmdVersion = "cmdVersion",
+    executable = "executable",
+    emulator = "emulator",
+    emulatorOpt = "emulatorOpt",
+    sdkManager = "sdkManager",
+}
 
-
+export enum ConfigScope {
+    global = 1,
+    workspace = 2,
+    folder = 3,
+    globalnWorkspace = 4,
+    globalnfolder = 5,
+}
 
 export class Manager {
     private static instance: Manager;
@@ -39,40 +52,79 @@ export class Manager {
         return Manager.instance;
     }
 
+    readonly android: AndroidService;
     readonly avd: AVDService;
     readonly sdk: SDKService;
-    readonly android: AndroidService;
     readonly output: Output;
     readonly cache: Cache;
 
     private constructor() {
         this.cache = new Cache();
+
+        this.android = new AndroidService(this);
         this.avd = new AVDService(this);
         this.sdk = new SDKService(this);
-        this.android = new AndroidService(this);
 
         this.output = new Output("AVD Manager");
     }
 
+    /**
+     * setConfig
+     * @param key string
+     * @param value any
+     * @param scope ConfigScope
+     */
+
+
+    public async setConfig(key: string, value: any, scope: ConfigScope = ConfigScope.folder) {
+        let config = workspace.getConfiguration('avdmanager');
+
+        if (scope === ConfigScope.global) {
+            await config.update(key, value, true);
+
+        } else if (scope === ConfigScope.workspace) {
+            await config.update(key, value, false);
+
+        } else if (scope === ConfigScope.folder) {
+            await config.update(key, value, undefined);
+
+        } else if (scope === ConfigScope.globalnWorkspace) {
+            await config.update(key, value, true);
+            await config.update(key, value, false);
+
+        } else if (scope === ConfigScope.globalnfolder) {
+            await config.update(key, value, true);
+            await config.update(key, value, undefined);
+
+        }
+    }
 
     public getConfig(): IConfig {
         let config = workspace.getConfiguration('avdmanager');
-        let sdkPath = config.get<string>("sdkPath", "avdmanager");
-        let cmdVersion = config.get<string>("cmdVersion", "avdmanager");
+        let sdkPath = config.get<string>(ConfigItem.sdkPath, process.env.ANDROID_SDK_ROOT ?? "");
+        let cmdVersion = config.get<string>(ConfigItem.cmdVersion, "latest");
+
         let cmdPath = path.join(sdkPath, "cmdline-tools", cmdVersion, "bin");
+        let buildToolPath = path.join(sdkPath, "build-tools");
+        let platformToolsPath = path.join(sdkPath, "platform-tools");
+        let emuPath = path.join(sdkPath, "emulator");
+
+        let executable = config.get<string>(ConfigItem.executable, "avdmanager");
+        let sdkManager = config.get<string>(ConfigItem.sdkManager, "sdkmanager");
+        let emulator = config.get<string>(ConfigItem.emulator, "emulator");
 
         return {
             sdkPath: sdkPath,
             cmdPath: cmdPath,
-            buildToolPath: path.join(sdkPath, "build-tools"),
-            platformToolsPath: path.join(sdkPath, "platform-tools"),
-            emuPath: path.join(sdkPath, "emulator"),
+            buildToolPath: buildToolPath,
+            platformToolsPath: platformToolsPath,
+            emuPath: emuPath,
 
             cmdVersion: cmdVersion,
-            executable: config.get<string>("executable", "avdmanager"),
-            emulator: config.get<string>("emulator", "emulator"),
-            emulatorOpt: config.get<string>("emulatorOpt", ""),
-            sdkManager: config.get<string>("sdkManager", "sdkManager")
+            executable: executable,
+            sdkManager: sdkManager,
+            emulator: emulator,
+            emulatorOpt: config.get<string>(ConfigItem.emulatorOpt, "")
         };
     }